Split test_serve_dashboard into two test suites
